% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteoHazard-package.R
\docType{package}
\name{meteoHazard-package}
\alias{meteoHazard}
\alias{meteoHazard-package}
\title{meteoHazard: Meteorological Hazard Predictions for Waste Management}
\description{
Turns meteorological data into management-relevant predictions and warnings for waste-management operations. Provides the Thermal Work Limit (TWL), the maximum sustainable metabolic rate (W/m²) for workers in thermally stressful environments, based on the methodology of Brake & Bates (2002) \doi{10.1080/104732202317070947}, alongside (planned) odour-dispersion and wind-blown-litter predictors. Part of the 'tidyWaste' family of packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vorpalvorpal/meteoHazard}
  \item Report bugs at \url{https://github.com/vorpalvorpal/meteoHazard/issues}
}

}
\author{
\strong{Maintainer}: R. Shannon \email{burrows_51mustard@icloud.com}

}
\keyword{internal}
